requirements.txt
numpy==1.21.6
opencv_python==4.6.0.66
pandas==1.3.5
Pillow==9.3.0
PyMuPDF==1.21.0
pytorch_lightning==1.7.7
PyYAML==6.0
torch==1.12.1
torchmetrics==0.9.3
torchvision==0.13.1
camelot-py==0.11.0
tabula-py==2.6.0
scikit-learn==1.0.2